pyproject.toml
[project]
name = "fastapi-foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
]
